extract generator customizations
